move test to correct test target
