+++ crates/walrus-simtest/tests/simtest_event_blob.rs
@@ -14,7 +14,6 @@ mod tests {
     };
 
     use rocksdb::Options as RocksdbOptions;
-    use sui_rpc_api::Client as RpcClient;
     use tokio::sync::RwLock;
     use typed_store::rocks::be_fix_int_ser;
     use walrus_core::{BlobId, test_utils};
@@ -214,48 +213,9 @@ mod tests {
         panic!("Timeout waiting for last certified event blob");
     }
 
-    /// Kills all the storage nodes in the cluster.
-    async fn kill_all_storage_nodes(node_ids: &[sui_simulator::task::NodeId]) {
-        let handle = sui_simulator::runtime::Handle::current();
-        for &node_id in node_ids {
-            handle.delete_node(node_id);
-        }
-    }
-
     async fn kill_node(node_id: sui_simulator::task::NodeId) {
         let handle = sui_simulator::runtime::Handle::current();
         handle.delete_node(node_id);
-    }
-
-    async fn restart_nodes_with_checkpoints(
-        walrus_cluster: &mut TestCluster<SimStorageNodeHandle>,
-        checkpoint_fn: impl Fn(usize) -> u32,
-    ) {
-        let node_handles = walrus_cluster
-            .nodes
-            .iter()
-            .map(|n| n.node_id.expect("simtest must set node id"))
-            .collect::<Vec<_>>();
-
-        kill_all_storage_nodes(&node_handles).await;
-
-        // Doing a short sleep so that the connection between the old nodes to the fullnodes are
-        // fully cleared. We've seen cases where when the new node is started, the old connection
-        // on the fullnode still exists and the new node will not be able to connect to the
-        //fullnode.
-        tokio::time::sleep(Duration::from_secs(2)).await;
-
-        for (i, node) in walrus_cluster.nodes.iter_mut().enumerate() {
-            node.node_id = Some(
-                SimStorageNodeHandle::spawn_node(
-                    Arc::new(RwLock::new(node.storage_node_config.clone())),
-                    Some(checkpoint_fn(i)),
-                    node.cancel_token.clone(),
-                )
-                .await
-                .id(),
-            );
-        }
     }
 
     async fn restart_node_with_checkpoints(
@@ -309,7 +269,7 @@ mod tests {
         tokio::time::sleep(Duration::from_secs(30)).await;
 
         // Restart nodes with different checkpoint numbers to create fork
-        restart_nodes_with_checkpoints(&mut walrus_cluster, |i| 30 + i as u32).await;
+        simtest_utils::restart_nodes_with_checkpoints(&mut walrus_cluster, |i| 30 + i as u32).await;
 
         // Wait for event blob certification to get stuck
         let stuck_blob = wait_for_certification_stuck(&client).await;
@@ -317,7 +277,7 @@ mod tests {
         tracing::info!("stuck blob: {:?}", stuck_blob);
 
         // Restart nodes with same checkpoint number to recover
-        restart_nodes_with_checkpoints(&mut walrus_cluster, |_| 20).await;
+        simtest_utils::restart_nodes_with_checkpoints(&mut walrus_cluster, |_| 20).await;
 
         // Verify recovery
         tokio::time::sleep(Duration::from_secs(40)).await;
@@ -376,138 +336,5 @@ mod tests {
         wait_for_event_blob_writer_to_recover(&walrus_cluster.nodes[0])
             .await
             .unwrap();
-    }
-
-    /// Waits for all nodes to download checkpoints up to the specified sequence number
-    async fn wait_for_nodes_at_checkpoint(
-        node_refs: &[&SimStorageNodeHandle],
-        target_sequence_number: u64,
-        timeout: Duration,
-    ) -> Result<(), anyhow::Error> {
-        let start_time = Instant::now();
-        let poll_interval = Duration::from_secs(1);
-        let mut nodes_to_check: Vec<&SimStorageNodeHandle> = node_refs.iter().copied().collect();
-
-        tracing::info!(
-            "Waiting for all nodes to download checkpoint up to sequence number: {}",
-            target_sequence_number
-        );
-
-        while !nodes_to_check.is_empty() {
-            if start_time.elapsed() > timeout {
-                return Err(anyhow::anyhow!(
-                    "Timeout waiting for nodes to download checkpoint.",
-                ));
-            }
-
-            let node_health_infos =
-                simtest_utils::get_nodes_health_info(nodes_to_check.clone()).await;
-
-            let lagging_nodes: Vec<&SimStorageNodeHandle> = node_health_infos
-                .iter()
-                .zip(nodes_to_check.iter())
-                .filter_map(
-                    |(info, node)| match info.latest_checkpoint_sequence_number {
-                        Some(seq) if seq < target_sequence_number => Some(*node),
-                        None => Some(*node),
-                        _ => None,
-                    },
-                )
-                .collect();
-
-            nodes_to_check = lagging_nodes;
-
-            tokio::time::sleep(poll_interval).await;
-        }
-
-        Ok(())
-    }
-
-    /// This test verifies that the node can correctly download checkpoints from
-    /// additional fullnodes.
-    #[ignore = "ignore integration simtests by default"]
-    #[walrus_simtest]
-    async fn test_checkpoint_downloader_with_additional_fullnodes() {
-        let checkpoints_per_event_blob = 20;
-        let (sui_cluster, mut walrus_cluster, client, _) = test_cluster::E2eTestSetupBuilder::new()
-            .with_epoch_duration(Duration::from_secs(15))
-            .with_test_nodes_config(TestNodesConfig {
-                node_weights: vec![2, 2, 3, 3, 3],
-                ..Default::default()
-            })
-            .with_num_checkpoints_per_blob(checkpoints_per_event_blob)
-            .with_communication_config(
-                ClientCommunicationConfig::default_for_test_with_reqwest_timeout(
-                    Duration::from_secs(2),
-                ),
-            )
-            .with_additional_fullnodes(4)
-            .build_generic::<SimStorageNodeHandle>()
-            .await
-            .unwrap();
-
-        // Register a fail point that will fail the first attempt.
-        sui_macros::register_fail_point_if("fallback_client_inject_error", move || true);
-        let primary_rpc_url = sui_cluster.lock().await.rpc_url();
-        let primary_rpc_url_clone = primary_rpc_url.clone();
-
-        // Always fail sui client creation for the primary rpc node.
-        sui_macros::register_fail_point_arg(
-            "failpoint_sui_client_build_client",
-            move || -> Option<String> { Some(primary_rpc_url.clone()) },
-        );
-
-        // Always fail rpc client creation for the primary rpc node.
-        sui_macros::register_fail_point_arg(
-            "failpoint_rpc_client_build_client",
-            move || -> Option<String> { Some(primary_rpc_url_clone.clone()) },
-        );
-
-        tracing::info!(
-            "Additional fullnodes: {:?}",
-            sui_cluster.lock().await.additional_rpc_urls()
-        );
-        let client_arc = Arc::new(client);
-
-        // Restart all nodes, this should still form a cluster with all nodes running.
-        restart_nodes_with_checkpoints(&mut walrus_cluster, |_| checkpoints_per_event_blob).await;
-
-        // Wait for the cluster to process some events.
-        let workload_handle =
-            simtest_utils::start_background_workload(client_arc.clone(), false, 0, None);
-        tokio::time::sleep(Duration::from_secs(30)).await;
-
-        // Get the latest checkpoint from Sui.
-        let rpc_client = RpcClient::new(sui_cluster.lock().await.additional_rpc_urls()[0].clone())
-            .expect("Failed to create RPC client");
-        let latest_sui_checkpoint = rpc_client
-            .get_latest_checkpoint()
-            .await
-            .expect("Failed to get latest checkpoint from Sui");
-
-        let latest_sui_checkpoint_seq = latest_sui_checkpoint.sequence_number;
-        tracing::info!(
-            "Latest Sui checkpoint sequence number: {}",
-            latest_sui_checkpoint_seq
-        );
-        workload_handle.abort();
-
-        // Get the highest processed event and checkpoint for each storage node.
-        let node_refs: Vec<&SimStorageNodeHandle> = walrus_cluster.nodes.iter().collect();
-        let node_health_infos = simtest_utils::get_nodes_health_info(node_refs.clone()).await;
-
-        tracing::info!("node health infos: {:?}", node_health_infos);
-
-        wait_for_nodes_at_checkpoint(
-            &node_refs,
-            latest_sui_checkpoint_seq,
-            Duration::from_secs(100),
-        )
-        .await
-        .expect("All nodes should have downloaded the checkpoint");
-
-        sui_macros::clear_fail_point("fallback_client_inject_error");
-        sui_macros::clear_fail_point("failpoint_sui_client_build_client");
-        sui_macros::clear_fail_point("failpoint_rpc_client_build_client");
     }
 }
